Fix for camera crash
- Added a conditional loop that will run until the Surface is available
for the Camera Preview to show.
-- On initial startup the app wasn't having any issues with creating the
Surface and then showing the Preview, but when exiting the Fragment and
then going back in, it was crashing because the Surface was null. Debug
breakpoints showed no issues which seemed to indicate that there might
be a race condition occurring. Adding in a while loop until the Surface
becomes available seems to have corrected this issue.
- Added a ProgressBar element to be shown when the Camera Preview isn't
immediately available.
-- On a fairly frequent basis, maybe one in four attempts, the Camera
Preview will take a few seconds to actually render on the screen so
an indeterminate ProgressBar was added to give some semblace of work
being done.

AD-1588

Change-Id: Ia5e7332f203d04c458854a2ee402f5e407f8e1ea