Button Demo Fix and Device Permission Update
- Fixed button image overlay issue in Button Demo.
- Added DeviceHelper to help facilitate device permission check and
requests, and updated this process in M400cActivity and M400cFragment.
-- A previous issue with Camera Demo needing all device permissions to
work properly means we need to front-load the checks and requests, but
the initial implementation didn't work to make sure all three devices
are accounted for. This update introduces a bit of recursion to ensure
this occurs. It's also structured in such a way that it could be added
to the SDK if we want.

AD-1641

Change-Id: If4fff997ba98576bb33971538f09321125fc9e28